Revert "Purchase Pack Desktop"
