+++ components/modals/TransactionModal.js
@@ -12,15 +12,20 @@ const TransactionModal = (props) => {
   const renderModalContent = () => {
 		if(modalStatus == statusCode.PENDING){
 			return (
-        <>
-          <LoadingModal/>
-        </>
+				<LoadingModal/>
 			)
 		}
 		if(modalStatus == statusCode.SUCCESS){
 			return (
 				<>
           {modalData.map((data) => `${data.name}: ${data.value}`)}
+          <button
+            type="button"
+            className="inline-flex justify-center px-4 py-2 text-sm font-medium text-blue-900 bg-blue-100 border border-transparent rounded-md hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
+            onClick={() => onClose()}
+          >
+            Close
+          </button>
         </>
 			)
 		}
@@ -28,18 +33,34 @@ const TransactionModal = (props) => {
 			return (
 				<>
           {modalData.map((data) => `${data.name}: ${data.value}`)}
+          <button
+            type="button"
+            className="inline-flex justify-center px-4 py-2 text-sm font-medium text-blue-900 bg-blue-100 border border-transparent rounded-md hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
+            onClick={() => onClose()}
+          >
+            Close
+          </button>
         </>
 			)
 		}
 		if(modalStatus == statusCode.ERROR){
 			return (
 				<>
-          {modalData.map((data) => `${data.value}`)}
+          {modalData.forEach((data) => {
+            return `${data.name}: ${data.value}`
+          })}
+          <button
+            type="button"
+            className="inline-flex justify-center px-4 py-2 text-sm font-medium text-blue-900 bg-blue-100 border border-transparent rounded-md hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
+            onClick={() => onClose()}
+          >
+            Close
+          </button>
         </>
 			)
 		}
 	}
-  //TODO: Make the modal more presentable
+
   return (
     <>
       <Transition appear show={visible} as={Fragment}>
@@ -85,13 +106,6 @@ const TransactionModal = (props) => {
                   {title}
                 </Dialog.Title>
                 {renderModalContent()}
-                <button
-                  type="button"
-                  className="inline-flex justify-center px-4 py-2 text-sm font-medium text-blue-900 bg-blue-100 border border-transparent rounded-md hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
-                  onClick={() => onClose()}
-                >
-                  Close
-                </button>
               </div>
             </Transition.Child>
           </div>
